Spark-36096: Group exception messages in resource
