+++ core/src/main/scala/org/apache/spark/resource/ResourceInformation.scala
@@ -22,9 +22,8 @@ import scala.util.control.NonFatal
 import org.json4s.{DefaultFormats, Extraction, JValue}
 import org.json4s.jackson.JsonMethods._
 
-import org.apache.spark.SparkException
 import org.apache.spark.annotation.Evolving
-import org.apache.spark.errors.ResourceErrors
+import org.apache.spark.errors.ParsingErrors
 
 /**
  * Class to hold information about a type of Resource. A resource could be a GPU, FPGA, etc.
@@ -73,7 +72,7 @@ private[spark] object ResourceInformation {
       parse(json).extract[ResourceInformationJson].toResourceInformation
     } catch {
       case NonFatal(e) =>
-        throw ResourceErrors.ErrorParseJson(json, exampleJson, e)
+        throw ParsingErrors.ErrorParseJson(json, exampleJson, e)
     }
   }
 
@@ -83,7 +82,7 @@ private[spark] object ResourceInformation {
       json.extract[ResourceInformationJson].toResourceInformation
     } catch {
       case NonFatal(e) =>
-        throw ResourceErrors.ErrorParseJson(json, e)
+        throw ParsingErrors.ErrorParseJson(json, e)
     }
   }
 }
